Move response_file and response_file_variables out of base package resource

Fix for Move response_file and response_file_variables out of base package resource


update require changes


Updated require changes


Fixed some unit test cases
