+++ spec/data/cookbooks/irssi/files/default/irssi.response
@@ -1,0 +1,2 @@
+# Hi, I'm pretending to be the preseed file for installing the irssi on debian
+# or Ubuntu
