+++ spec/data/cookbooks/wget/files/default/wget.response
@@ -1,0 +1,2 @@
+# Hi, I'm pretending to be the preseed file for installing the wget on debian
+# or Ubuntu
