+++ spec/unit/resource/dpkg_package_spec.rb
@@ -36,6 +36,16 @@ describe Chef::Resource::DpkgPackage, "initialize" do
       expect(resource.action).to eql([:install])
     end
 
+    it "accepts a string for the response file" do
+      resource.response_file "something"
+      expect(resource.response_file).to eql("something")
+    end
+
+    it "accepts a hash for response file template variables" do
+      resource.response_file_variables({ variables: true })
+      expect(resource.response_file_variables).to eql({ variables: true })
+    end
+
     it "supports :install, :lock, :purge, :reconfig, :remove, :unlock, :upgrade actions" do
       expect { resource.action :install }.not_to raise_error
       expect { resource.action :lock }.not_to raise_error
